Refactor upload_packs.py added test_upload_packs_private.py
