+++ Tests/Marketplace/Tests/upload_packs_test.py
@@ -390,58 +390,6 @@ class TestUpdateIndex:
             assert call_arg[0] in expected_copy_args
 
 
-class TestPrivatePacks:
-    def test_add_private_packs_to_index(self, mocker):
-        from Tests.Marketplace import upload_packs
-
-        dirs = scan_dir()
-        mocker.patch('os.scandir', return_value=dirs)
-        mocker.patch('os.path.isdir', side_effect=FakeDirEntry.isdir)
-        mocker.patch.object(upload_packs, 'update_index_folder')
-
-        upload_packs.add_private_packs_to_index('test', 'private_test')
-
-        index_call_args = upload_packs.update_index_folder.call_args[0]
-        index_call_count = upload_packs.update_index_folder.call_count
-
-        assert index_call_count == 1
-        assert index_call_args[0] == 'test'
-        assert index_call_args[1] == 'mock_dir'
-        assert index_call_args[2] == 'mock_path'
-
-    def test_get_private_packs(self, mocker):
-        import os
-        from Tests.Marketplace import upload_packs, marketplace_services
-
-        mocker.patch('glob.glob', return_value=[os.path.join(marketplace_services.CONTENT_ROOT_PATH,
-                                                             'Tests', 'Marketplace', 'Tests',
-                                                             'test_data', 'metadata.json')])
-
-        private_packs = upload_packs.get_private_packs('path')
-
-        assert private_packs == [{'id': 'ImpossibleTraveler', 'price': 100}]
-
-    def test_get_private_packs_empty(self, mocker):
-        from Tests.Marketplace import upload_packs
-
-        mocker.patch('glob.glob', return_value=[])
-        mocker.patch("Tests.Marketplace.upload_packs.logging.warning")
-
-        private_packs = upload_packs.get_private_packs('path')
-
-        assert private_packs == []
-
-    def test_get_private_packs_error(self, mocker):
-        from Tests.Marketplace import upload_packs
-
-        mocker.patch('glob.glob', side_effect=InterruptedError)
-        mocker.patch("Tests.Marketplace.upload_packs.logging.warning")
-
-        private_packs = upload_packs.get_private_packs('path')
-
-        assert private_packs == []
-
-
 class TestCleanPacks:
     """ Test for clean_non_existing_packs function scenarios.
     """
