+++ Tests/Marketplace/upload_packs.py
@@ -8,19 +8,22 @@ import prettytable
 import glob
 import git
 import requests
+import logging
 from datetime import datetime
 from zipfile import ZipFile
-import logging
-
-from Tests.scripts.utils.log_util import install_logging
+from typing import Any, Tuple, Union
 from Tests.Marketplace.marketplace_services import init_storage_client, init_bigquery_client, Pack, PackStatus, \
     GCPConfig, PACKS_FULL_PATH, IGNORED_FILES, PACKS_FOLDER, IGNORED_PATHS, Metadata, CONTENT_ROOT_PATH, \
     get_packs_statistics_dataframe
+from Tests.private_build.upload_packs_private import get_private_packs, add_private_packs_to_index, update_index_with_priced_packs
 from demisto_sdk.commands.common.tools import run_command, str2bool
 
+from Tests.scripts.utils.log_util import install_logging
 
-def get_packs_names(target_packs):
-    """Detects and returns packs names to upload.
+
+def get_packs_names(target_packs: str) -> set:
+    """
+    Detects and returns packs names to upload.
 
     In case that `Modified` is passed in target_packs input, checks the git difference between two commits,
     current and previous and greps only ones with prefix Packs/.
@@ -41,7 +44,7 @@ def get_packs_names(target_packs):
             # return all available packs names
             return all_packs
         else:
-            logging.critical(f"Folder {PACKS_FOLDER} was not found at the following path: {PACKS_FULL_PATH}")
+            logging.error((f"Folder {PACKS_FOLDER} was not found at the following path: {PACKS_FULL_PATH}"))
             sys.exit(1)
     elif target_packs.lower() == "modified":
         cmd = "git diff --name-only HEAD..HEAD^ | grep 'Packs/'"
@@ -60,7 +63,7 @@ def get_packs_names(target_packs):
         sys.exit(1)
 
 
-def extract_packs_artifacts(packs_artifacts_path, extract_destination_path):
+def extract_packs_artifacts(packs_artifacts_path: str, extract_destination_path: str):
     """Extracts all packs from content pack artifact zip.
 
     Args:
@@ -73,7 +76,7 @@ def extract_packs_artifacts(packs_artifacts_path, extract_destination_path):
     logging.info("Finished extracting packs artifacts")
 
 
-def download_and_extract_index(storage_bucket, extract_destination_path):
+def download_and_extract_index(storage_bucket: Any, extract_destination_path: str) -> Tuple[str, Any, int]:
     """Downloads and extracts index zip from cloud storage.
 
     Args:
@@ -96,11 +99,13 @@ def download_and_extract_index(storage_bucket, extract_destination_path):
         os.mkdir(extract_destination_path)
 
     if not index_blob.exists():
+        logging.error("The blob does not exist.")
         os.mkdir(index_folder_path)
         return index_folder_path, index_blob, index_generation
 
     index_blob.reload()
     index_generation = index_blob.generation
+
     index_blob.download_to_filename(download_index_path, if_generation_match=index_generation)
 
     if os.path.exists(download_index_path):
@@ -120,8 +125,10 @@ def download_and_extract_index(storage_bucket, extract_destination_path):
         sys.exit(1)
 
 
-def update_index_folder(index_folder_path, pack_name, pack_path, pack_version='', hidden_pack=False):
-    """Copies pack folder into index folder.
+def update_index_folder(index_folder_path: str, pack_name: str, pack_path: str, pack_version: str = '',
+                        hidden_pack: bool = False) -> bool:
+    """
+    Copies pack folder into index folder.
 
     Args:
         index_folder_path (str): full path to index folder.
@@ -158,8 +165,7 @@ def update_index_folder(index_folder_path, pack_name, pack_path, pack_version=''
             if os.path.exists(index_pack_path):
                 shutil.rmtree(index_pack_path)  # remove pack folder inside index in case that it exists
             logging.warning(f"Skipping updating {pack_name} pack files to index")
-            task_status = True
-            return
+            return True
 
         # Copy new files and add metadata for latest version
         for d in os.scandir(pack_path):
@@ -173,12 +179,12 @@ def update_index_folder(index_folder_path, pack_name, pack_path, pack_version=''
 
         task_status = True
     except Exception:
-        logging.exception(f"Failed in updating index folder for {pack_name} pack\n")
+        logging.exception(f"Failed in updating index folder for {pack_name} pack.")
     finally:
         return task_status
 
 
-def clean_non_existing_packs(index_folder_path, private_packs, storage_bucket):
+def clean_non_existing_packs(index_folder_path: str, private_packs: list, storage_bucket: Any) -> bool:
     """ Detects packs that are not part of content repo or from private packs bucket.
 
     In case such packs were detected, problematic pack is deleted from index and from content/packs/{target_pack} path.
@@ -229,18 +235,21 @@ def clean_non_existing_packs(index_folder_path, private_packs, storage_bucket):
     return False
 
 
-def upload_index_to_storage(index_folder_path, extract_destination_path, index_blob, build_number, private_packs,
-                            current_commit_hash, index_generation):
-    """Upload updated index zip to cloud storage.
+def upload_index_to_storage(index_folder_path: str, extract_destination_path: str, index_blob: Any,
+                            build_number: str, private_packs: list, current_commit_hash: str,
+                            index_generation: str, is_private: bool = False):
+    """
+    Upload updated index zip to cloud storage.
 
-    Args:
-        index_folder_path (str): index folder full path.
-        extract_destination_path (str): extract folder full path.
-        index_blob (Blob): google cloud storage object that represents index.zip blob.
-        build_number (str): circleCI build number, used as an index revision.
-        private_packs (list): List of private packs and their price.
-        current_commit_hash (str): last commit hash of head.
-        index_generation (str): downloaded index generation.
+    :param index_folder_path: index folder full path.
+    :param extract_destination_path: extract folder full path.
+    :param index_blob: google cloud storage object that represents index.zip blob.
+    :param build_number: circleCI build number, used as an index revision.
+    :param private_packs: List of private packs and their price.
+    :param current_commit_hash: last commit hash of head.
+    :param index_generation: downloaded index generation.
+    :param is_private: Indicates if upload is private.
+    :returns None.
 
     """
     with open(os.path.join(index_folder_path, f"{GCPConfig.INDEX_NAME}.json"), "w+") as index_file:
@@ -260,22 +269,22 @@ def upload_index_to_storage(index_folder_path, extract_destination_path, index_b
         current_index_generation = index_blob.generation
         index_blob.cache_control = "no-cache,max-age=0"  # disabling caching for index blob
 
-        if current_index_generation == index_generation:
+        if is_private or current_index_generation == index_generation:
             index_blob.upload_from_filename(index_zip_path)
             logging.success(f"Finished uploading {GCPConfig.INDEX_NAME}.zip to storage.")
         else:
-            logging.error(f"Failed in uploading {GCPConfig.INDEX_NAME}, mismatch in index file generation")
-            logging.error(f"Downloaded index generation: {index_generation}")
-            logging.error(f"Current index generation: {current_index_generation}")
+            logging.critical(f"Failed in uploading {GCPConfig.INDEX_NAME}, mismatch in index file generation")
+            logging.critical(f"Downloaded index generation: {index_generation}")
+            logging.critical(f"Current index generation: {current_index_generation}")
             sys.exit(0)
     except Exception:
-        logging.exception(f"Failed in uploading {GCPConfig.INDEX_NAME}")
+        logging.exception(f"Failed in uploading {GCPConfig.INDEX_NAME}.")
         sys.exit(1)
     finally:
         shutil.rmtree(index_folder_path)
 
 
-def upload_core_packs_config(storage_bucket, build_number, index_folder_path):
+def upload_core_packs_config(storage_bucket: Any, build_number: str, index_folder_path: str):
     """Uploads corepacks.json file configuration to bucket. Corepacks file includes core packs for server installation.
 
      Args:
@@ -286,7 +295,6 @@ def upload_core_packs_config(storage_bucket, build_number, index_folder_path):
     """
     core_packs_public_urls = []
     found_core_packs = set()
-
     for pack in os.scandir(index_folder_path):
         if pack.is_dir() and pack.name in GCPConfig.CORE_PACKS_LIST:
             pack_metadata_path = os.path.join(index_folder_path, pack.name, Pack.METADATA)
@@ -312,8 +320,8 @@ def upload_core_packs_config(storage_bucket, build_number, index_folder_path):
 
     if len(found_core_packs) != len(GCPConfig.CORE_PACKS_LIST):
         missing_core_packs = set(GCPConfig.CORE_PACKS_LIST) ^ found_core_packs
-        logging.error(f"Number of defined core packs are: {len(GCPConfig.CORE_PACKS_LIST)}")
-        logging.error(f"Actual number of found core packs are: {len(found_core_packs)}")
+        logging.critical(f"Number of defined core packs are: {len(GCPConfig.CORE_PACKS_LIST)}")
+        logging.critical(f"Actual number of found core packs are: {len(found_core_packs)}")
         logging.critical(f"Missing core packs are: {missing_core_packs}")
         sys.exit(1)
 
@@ -330,7 +338,7 @@ def upload_core_packs_config(storage_bucket, build_number, index_folder_path):
     logging.success(f"Finished uploading {GCPConfig.CORE_PACK_FILE_NAME} to storage.")
 
 
-def upload_id_set(storage_bucket, id_set_local_path=None):
+def upload_id_set(storage_bucket: Any, id_set_local_path: str = None):
     """
     Uploads the id_set.json artifact to the bucket.
 
@@ -350,91 +358,7 @@ def upload_id_set(storage_bucket, id_set_local_path=None):
     logging.success("Finished uploading id_set.json to storage.")
 
 
-def get_private_packs(private_index_path):
-    """ Get the list of ID and price of the private packs.
-
-    Args:
-        private_index_path: The path for the index of the private packs.
-
-    Returns:
-        private_packs: A list of ID and price of the private packs.
-    """
-    try:
-        metadata_files = glob.glob(f"{private_index_path}/**/metadata.json")
-    except Exception as e:
-        logging.warning(f'Could not find metadata files in {private_index_path}: {str(e)}')
-        return []
-
-    if not metadata_files:
-        logging.warning(f'No metadata files found in [{private_index_path}]')
-
-    private_packs = []
-    for metadata_file_path in metadata_files:
-        try:
-            with open(metadata_file_path, "r") as metadata_file:
-                metadata = json.load(metadata_file)
-            if metadata:
-                private_pack = {
-                    'id': metadata.get('id'),
-                    'price': metadata.get('price'),
-                }
-
-                if metadata.get('vendorId') is not None:
-                    private_pack['vendorId'] = metadata.get('vendorId')
-                if metadata.get('vendorName') is not None:
-                    private_pack['vendorName'] = metadata.get('vendorName')
-
-                private_packs.append(private_pack)
-
-        except ValueError:
-            logging.exception(f'Invalid JSON in the metadata file [{metadata_file_path}].')
-
-    return private_packs
-
-
-def add_private_packs_to_index(index_folder_path, private_index_path):
-    """ Add the private packs to the index folder.
-
-    Args:
-        index_folder_path: The index folder path.
-        private_index_path: The path for the index of the private packs.
-
-    """
-    for d in os.scandir(private_index_path):
-        if os.path.isdir(d.path):
-            update_index_folder(index_folder_path, d.name, d.path)
-
-
-def update_index_with_priced_packs(private_storage_bucket, extract_destination_path, index_folder_path):
-    """ Updates index with priced packs and returns list of priced packs data.
-
-    Args:
-        private_storage_bucket (google.cloud.storage.bucket.Bucket): google storage private bucket.
-        extract_destination_path (str): full path to extract directory.
-        index_folder_path (str): downloaded index folder directory path.
-
-    Returns:
-        list: priced packs from private bucket.
-
-    """
-    private_index_path = ""
-    private_packs = []
-
-    try:
-        private_index_path, _, _ = download_and_extract_index(private_storage_bucket,
-                                                              os.path.join(extract_destination_path, 'private'))
-        private_packs = get_private_packs(private_index_path)
-        add_private_packs_to_index(index_folder_path, private_index_path)
-        logging.info("Finished updating index with priced packs")
-    except Exception:
-        logging.exception('Could not add private packs to the index.')
-    finally:
-        if private_index_path:
-            shutil.rmtree(os.path.dirname(private_index_path), ignore_errors=True)
-        return private_packs
-
-
-def _build_summary_table(packs_input_list, include_pack_status=False):
+def _build_summary_table(packs_input_list: list, include_pack_status: bool = False) -> Any:
     """Build summary table from pack list
 
     Args:
@@ -460,7 +384,7 @@ def _build_summary_table(packs_input_list, include_pack_status=False):
     return table
 
 
-def build_summary_table_md(packs_input_list, include_pack_status=False):
+def build_summary_table_md(packs_input_list: list, include_pack_status: bool = False) -> str:
     """Build markdown summary table from pack list
 
     Args:
@@ -471,8 +395,9 @@ def build_summary_table_md(packs_input_list, include_pack_status=False):
         Markdown table: table with upload result of packs.
 
     """
-    table_fields = ["Index", "Pack ID", "Pack Display Name", "Latest Version", "Status"] if include_pack_status \
-        else ["Index", "Pack ID", "Pack Display Name", "Latest Version"]
+    table_fields = ["Index", "Pack ID", "Pack Display Name", "Latest Version", "Status",
+                    "Pack Bucket URL"] if include_pack_status \
+        else ["Index", "Pack ID", "Pack Display Name", "Latest Version", "Pack Bucket URL"]
 
     table = ['|', '|']
 
@@ -483,8 +408,9 @@ def build_summary_table_md(packs_input_list, include_pack_status=False):
     for index, pack in enumerate(packs_input_list):
         pack_status_message = PackStatus[pack.status].value if include_pack_status else ''
 
-        row = [index, pack.name, pack.display_name, pack.latest_version, pack_status_message] if include_pack_status \
-            else [index, pack.name, pack.display_name, pack.latest_version]
+        row = [index, pack.name, pack.display_name, pack.latest_version, pack_status_message,
+               pack.bucket_url] if include_pack_status \
+            else [index, pack.name, pack.display_name, pack.latest_version, pack.bucket_url]
 
         row_hr = '|'
         for _value in row:
@@ -494,7 +420,7 @@ def build_summary_table_md(packs_input_list, include_pack_status=False):
     return '\n'.join(table)
 
 
-def load_json(file_path):
+def load_json(file_path: str) -> dict:
     """ Reads and loads json file.
 
     Args:
@@ -504,13 +430,15 @@ def load_json(file_path):
         dict: loaded json file.
 
     """
-    with open(file_path, 'r') as json_file:
-        result = json.load(json_file)
-
+    if file_path:
+        with open(file_path, 'r') as json_file:
+            result = json.load(json_file)
+    else:
+        result = {}
     return result
 
 
-def get_content_git_client(content_repo_path):
+def get_content_git_client(content_repo_path: str):
     """ Initializes content repo client.
 
     Args:
@@ -523,7 +451,7 @@ def get_content_git_client(content_repo_path):
     return git.Repo(content_repo_path)
 
 
-def get_recent_commits_data(content_repo):
+def get_recent_commits_data(content_repo: Any):
     """ Returns recent commits hashes (of head and remote master)
 
     Args:
@@ -536,10 +464,10 @@ def get_recent_commits_data(content_repo):
     return content_repo.head.commit.hexsha, content_repo.commit('origin/master~1').hexsha
 
 
-def check_if_index_is_updated(index_folder_path, content_repo, current_commit_hash, remote_previous_commit_hash,
-                              storage_bucket):
-    """ Checks stored at index.json commit hash and compares it to current commit hash. In case no packs folders were
-    added/modified/deleted, all other steps are not performed.
+def check_if_index_is_updated(index_folder_path: str, content_repo: Any, current_commit_hash: str,
+                              remote_previous_commit_hash: str, storage_bucket: Any):
+    """ Checks stored at index.json commit hash and compares it to current commit hash. In case no
+    packs folders were added/modified/deleted, all other steps are not performed.
 
     Args:
         index_folder_path (str): index folder full path.
@@ -568,17 +496,15 @@ def check_if_index_is_updated(index_folder_path, content_repo, current_commit_ha
 
         try:
             index_commit = content_repo.commit(index_commit_hash)
-        except Exception as e:
+        except Exception:
             # not updated build will receive this exception because it is missing more updated commit
-            logging.warning(f"Index is already updated. Additional info:\n {e}")
-            logging.warning(skipping_build_task_message)
+            logging.exception(f"Index is already updated. {skipping_build_task_message}")
             sys.exit()
 
         current_commit = content_repo.commit(current_commit_hash)
 
         if current_commit.committed_datetime <= index_commit.committed_datetime:
-            logging.warning(
-                f"Current commit {current_commit.hexsha} committed time: {current_commit.committed_datetime}")
+            logging.warning(f"Current commit {current_commit.hexsha} committed time: {current_commit.committed_datetime}")
             logging.warning(f"Index commit {index_commit.hexsha} committed time: {index_commit.committed_datetime}")
             logging.warning("Index is already updated.")
             logging.warning(skipping_build_task_message)
@@ -586,8 +512,7 @@ def check_if_index_is_updated(index_folder_path, content_repo, current_commit_ha
 
         for changed_file in current_commit.diff(index_commit):
             if changed_file.a_path.startswith(PACKS_FOLDER):
-                logging.info(
-                    f"Found changed packs between index commit {index_commit.hexsha} and {current_commit.hexsha}")
+                logging.info(f"Found changed packs between index commit {index_commit.hexsha} and {current_commit.hexsha}")
                 break
         else:
             logging.warning(f"No changes found between index commit {index_commit.hexsha} and {current_commit.hexsha}")
@@ -598,7 +523,7 @@ def check_if_index_is_updated(index_folder_path, content_repo, current_commit_ha
         sys.exit(1)
 
 
-def print_packs_summary(packs_list):
+def print_packs_summary(packs_list: list):
     """Prints summary of packs uploaded to gcs.
 
     Args:
@@ -622,6 +547,8 @@ Total number of packs: {len(packs_list)}
         successful_packs_table = _build_summary_table(successful_packs)
         logging.success(f"Number of successful uploaded packs: {len(successful_packs)}")
         logging.success(f"Uploaded packs:\n{successful_packs_table}")
+        with open('pack_list.txt', 'w') as f:
+            f.write(successful_packs_table.get_string())
     if skipped_packs:
         skipped_packs_table = _build_summary_table(skipped_packs)
         logging.warning(f"Number of skipped packs: {len(skipped_packs)}")
@@ -633,18 +560,18 @@ Total number of packs: {len(packs_list)}
         sys.exit(1)
 
     # for external pull requests -  when there is no failed packs, add the build summary to the pull request
-    branch_name = os.environ['CIRCLE_BRANCH']
-    if branch_name.startswith('pull/'):
+    branch_name = os.environ.get('CIRCLE_BRANCH')
+    if branch_name and branch_name.startswith('pull/'):
         successful_packs_table = build_summary_table_md(successful_packs)
 
         build_num = os.environ['CIRCLE_BUILD_NUM']
 
         bucket_path = f'https://console.cloud.google.com/storage/browser/' \
-                      f'marketplace-ci-build/content/builds/{branch_name}/{build_num}'
+            f'marketplace-ci-build/content/builds/{branch_name}/{build_num}'
 
         pr_comment = f'Number of successful uploaded packs: {len(successful_packs)}\n' \
-                     f'Uploaded packs:\n{successful_packs_table}\n\n' \
-                     f'Browse to the build bucket with this address:\n{bucket_path}'
+            f'Uploaded packs:\n{successful_packs_table}\n\n' \
+            f'Browse to the build bucket with this address:\n{bucket_path}'
 
         add_pr_comment(pr_comment)
 
@@ -691,7 +618,7 @@ def option_handler():
     return parser.parse_args()
 
 
-def add_pr_comment(comment):
+def add_pr_comment(comment: str):
     """Add comment to the pull request.
 
     Args:
@@ -717,13 +644,18 @@ def add_pr_comment(comment):
             logging.warning(
                 f'Add pull request comment failed: There is more then one open pull request for branch {branch_name}.')
     except Exception:
-        logging.exception('Add pull request comment failed')
+        logging.exception('Add pull request comment failed.')
 
 
-def handle_github_response(response):
+def handle_github_response(response: json) -> dict:
+    """
+    Handles the response from the GitHub server after making a request.
+    :param response: Response from the server.
+    :return: The returned response.
+    """
     res_dict = response.json()
     if not res_dict.get('ok'):
-        logging.warning('Add pull request comment failed: {}'.format(res_dict.get('message')))
+        logging.warning(f'Add pull request comment failed: {res_dict.get("message")}')
     return res_dict
