+++ Tests/private_build/upload_packs_private.py
@@ -15,51 +15,13 @@ from typing import Any, Tuple, Union
 from Tests.Marketplace.marketplace_services import init_storage_client, init_bigquery_client, Pack, PackStatus, \
     GCPConfig, PACKS_FULL_PATH, IGNORED_FILES, PACKS_FOLDER, IGNORED_PATHS, Metadata, CONTENT_ROOT_PATH, \
     get_packs_statistics_dataframe
+from Tests.Marketplace.upload_packs import get_packs_names, extract_packs_artifacts, download_and_extract_index,\
+    update_index_folder, clean_non_existing_packs, upload_index_to_storage, upload_core_packs_config,\
+    upload_id_set, _build_summary_table, build_summary_table_md, load_json, get_content_git_client,\
+    get_recent_commits_data, check_if_index_is_updated, print_packs_summary, add_pr_comment, handle_github_response
 from demisto_sdk.commands.common.tools import run_command, str2bool
 
 from Tests.scripts.utils.log_util import install_logging
-
-
-def get_packs_names(target_packs: str) -> set:
-    """
-    Detects and returns packs names to upload.
-
-    In case that `Modified` is passed in target_packs input, checks the git difference between two commits,
-    current and previous and greps only ones with prefix Packs/.
-    By default this function will receive `All` as target_packs and will return all packs names from content repo.
-
-    Args:
-        target_packs (str): csv packs names or `All` for all available packs in content
-                            or `Modified` for only modified packs (currently not in use).
-
-    Returns:
-        set: unique collection of packs names to upload.
-
-    """
-    if target_packs.lower() == "all":
-        if os.path.exists(PACKS_FULL_PATH):
-            all_packs = {p for p in os.listdir(PACKS_FULL_PATH) if p not in IGNORED_FILES}
-            logging.info(f"Number of selected packs to upload is: {len(all_packs)}")
-            # return all available packs names
-            return all_packs
-        else:
-            logging.error((f"Folder {PACKS_FOLDER} was not found at the following path: {PACKS_FULL_PATH}"))
-            sys.exit(1)
-    elif target_packs.lower() == "modified":
-        cmd = "git diff --name-only HEAD..HEAD^ | grep 'Packs/'"
-        modified_packs_path = run_command(cmd).splitlines()
-        modified_packs = {p.split('/')[1] for p in modified_packs_path if p not in IGNORED_PATHS}
-        logging.info(f"Number of modified packs is: {len(modified_packs)}")
-        # return only modified packs between two commits
-        return modified_packs
-    elif target_packs and isinstance(target_packs, str):
-        modified_packs = {p.strip() for p in target_packs.split(',') if p not in IGNORED_FILES}
-        logging.info(f"Number of selected packs to upload is: {len(modified_packs)}")
-        # return only packs from csv list
-        return modified_packs
-    else:
-        logging.critical("Not correct usage of flag -p. Please check help section of upload packs script.")
-        sys.exit(1)
 
 
 def is_pack_paid_or_premium(path_to_pack_metadata_in_index: str) -> bool:
@@ -93,68 +55,6 @@ def delete_public_packs_from_index(index_folder_path: str):
             shutil.rmtree(path_to_pack, ignore_errors=True)
 
 
-def extract_packs_artifacts(packs_artifacts_path: str, extract_destination_path: str):
-    """Extracts all packs from content pack artifact zip.
-
-    Args:
-        packs_artifacts_path (str): full path to content artifacts zip file.
-        extract_destination_path (str): full path to directory where to extract the packs.
-
-    """
-    with ZipFile(packs_artifacts_path) as packs_artifacts:
-        packs_artifacts.extractall(extract_destination_path)
-    logging.info("Finished extracting packs artifacts")
-
-
-def download_and_extract_index(storage_bucket: Any, extract_destination_path: str) -> Tuple[str, Any, int]:
-    """Downloads and extracts index zip from cloud storage.
-
-    Args:
-        storage_bucket (google.cloud.storage.bucket.Bucket): google storage bucket where index.zip is stored.
-        extract_destination_path (str): the full path of extract folder.
-    Returns:
-        str: extracted index folder full path.
-        Blob: google cloud storage object that represents index.zip blob.
-        str: downloaded index generation.
-
-    """
-    index_storage_path = os.path.join(GCPConfig.STORAGE_BASE_PATH, f"{GCPConfig.INDEX_NAME}.zip")
-    download_index_path = os.path.join(extract_destination_path, f"{GCPConfig.INDEX_NAME}.zip")
-
-    index_blob = storage_bucket.blob(index_storage_path)
-    index_folder_path = os.path.join(extract_destination_path, GCPConfig.INDEX_NAME)
-    index_generation = 0  # Setting to 0 makes the operation succeed only if there are no live versions of the blob
-
-    if not os.path.exists(extract_destination_path):
-        os.mkdir(extract_destination_path)
-
-    if not index_blob.exists():
-        logging.error("The blob does not exist.")
-        os.mkdir(index_folder_path)
-        return index_folder_path, index_blob, index_generation
-
-    index_blob.reload()
-    index_generation = index_blob.generation
-
-    index_blob.download_to_filename(download_index_path, if_generation_match=index_generation)
-
-    if os.path.exists(download_index_path):
-        with ZipFile(download_index_path, 'r') as index_zip:
-            index_zip.extractall(extract_destination_path)
-
-        if not os.path.exists(index_folder_path):
-            logging.critical(f"Failed creating {GCPConfig.INDEX_NAME} folder with extracted data.")
-            sys.exit(1)
-
-        os.remove(download_index_path)
-        logging.info(f"Finished downloading and extracting {GCPConfig.INDEX_NAME} file to {extract_destination_path}")
-
-        return index_folder_path, index_blob, index_generation
-    else:
-        logging.critical(f"Failed to download {GCPConfig.INDEX_NAME}.zip file from cloud storage.")
-        sys.exit(1)
-
-
 def update_private_index(private_index_path: str, unified_index_path: str):
     """
     Updates the private index by copying the unified index to the private index.
@@ -170,239 +70,6 @@ def update_private_index(private_index_path: str, unified_index_path: str):
         path_to_pack_on_private_index = os.path.join(unified_index_path, private_pack_name)
         path_to_pack_on_unified_index = os.path.join(unified_index_path, private_pack_name)
         shutil.copy(path_to_pack_on_unified_index, path_to_pack_on_private_index)
-
-
-def update_index_folder(index_folder_path: str, pack_name: str, pack_path: str, pack_version: str = '',
-                        hidden_pack: bool = False) -> bool:
-    """
-    Copies pack folder into index folder.
-
-    Args:
-        index_folder_path (str): full path to index folder.
-        pack_name (str): pack folder name to copy.
-        pack_path (str): pack folder full path.
-        pack_version (str): pack latest version.
-        hidden_pack (bool): whether pack is hidden/internal or regular pack.
-
-    Returns:
-        bool: whether the operation succeeded.
-    """
-    task_status = False
-
-    try:
-        index_folder_subdirectories = [d for d in os.listdir(index_folder_path) if
-                                       os.path.isdir(os.path.join(index_folder_path, d))]
-        index_pack_path = os.path.join(index_folder_path, pack_name)
-        metadata_files_in_index = glob.glob(f"{index_pack_path}/metadata-*.json")
-        new_metadata_path = os.path.join(index_pack_path, f"metadata-{pack_version}.json")
-
-        if pack_version:
-            # Update the latest metadata
-            if new_metadata_path in metadata_files_in_index:
-                metadata_files_in_index.remove(new_metadata_path)
-
-        # Remove old files but keep metadata files
-        if pack_name in index_folder_subdirectories:
-            for d in os.scandir(index_pack_path):
-                if d.path not in metadata_files_in_index:
-                    os.remove(d.path)
-
-        # skipping index update in case hidden is set to True
-        if hidden_pack:
-            if os.path.exists(index_pack_path):
-                shutil.rmtree(index_pack_path)  # remove pack folder inside index in case that it exists
-            logging.warning(f"Skipping updating {pack_name} pack files to index")
-            return True
-
-        # Copy new files and add metadata for latest version
-        for d in os.scandir(pack_path):
-            if not os.path.exists(index_pack_path):
-                os.mkdir(index_pack_path)
-                logging.info(f"Created {pack_name} pack folder in {GCPConfig.INDEX_NAME}")
-
-            shutil.copy(d.path, index_pack_path)
-            if pack_version and Pack.METADATA == d.name:
-                shutil.copy(d.path, new_metadata_path)
-
-        task_status = True
-    except Exception:
-        logging.exception(f"Failed in updating index folder for {pack_name} pack.")
-    finally:
-        return task_status
-
-
-def clean_non_existing_packs(index_folder_path: str, private_packs: list, storage_bucket: Any) -> bool:
-    """ Detects packs that are not part of content repo or from private packs bucket.
-
-    In case such packs were detected, problematic pack is deleted from index and from content/packs/{target_pack} path.
-
-    Args:
-        index_folder_path (str): full path to downloaded index folder.
-        private_packs (list): priced packs from private bucket.
-        storage_bucket (google.cloud.storage.bucket.Bucket): google storage bucket where index.zip is stored.
-
-    Returns:
-        bool: whether cleanup was skipped or not.
-    """
-    if ('CI' not in os.environ) or (
-            os.environ.get('CIRCLE_BRANCH') != 'master' and storage_bucket.name == GCPConfig.PRODUCTION_BUCKET) or (
-            os.environ.get('CIRCLE_BRANCH') == 'master' and storage_bucket.name != GCPConfig.PRODUCTION_BUCKET):
-        logging.info("Skipping cleanup of packs in gcs.")  # skipping execution of cleanup in gcs bucket
-        return True
-
-    public_packs_names = {p for p in os.listdir(PACKS_FULL_PATH) if p not in IGNORED_FILES}
-    private_packs_names = {p.get('id', '') for p in private_packs}
-    valid_packs_names = public_packs_names.union(private_packs_names)
-    # search for invalid packs folder inside index
-    invalid_packs_names = {(entry.name, entry.path) for entry in os.scandir(index_folder_path) if
-                           entry.name not in valid_packs_names and entry.is_dir()}
-
-    if invalid_packs_names:
-        try:
-            logging.warning(f"Detected {len(invalid_packs_names)} non existing pack inside index, starting cleanup.")
-
-            for invalid_pack in invalid_packs_names:
-                invalid_pack_name = invalid_pack[0]
-                invalid_pack_path = invalid_pack[1]
-                # remove pack from index
-                shutil.rmtree(invalid_pack_path)
-                logging.warning(f"Deleted {invalid_pack_name} pack from {GCPConfig.INDEX_NAME} folder")
-                # important to add trailing slash at the end of path in order to avoid packs with same prefix
-                invalid_pack_gcs_path = os.path.join(GCPConfig.STORAGE_BASE_PATH, invalid_pack_name, "")  # by design
-
-                for invalid_blob in [b for b in storage_bucket.list_blobs(prefix=invalid_pack_gcs_path)]:
-                    logging.warning(f"Deleted invalid {invalid_pack_name} pack under url {invalid_blob.public_url}")
-                    invalid_blob.delete()  # delete invalid pack in gcs
-        except Exception:
-            logging.exception("Failed to cleanup non existing packs.")
-
-    else:
-        logging.info(f"No invalid packs detected inside {GCPConfig.INDEX_NAME} folder")
-
-    return False
-
-
-def upload_index_to_storage(index_folder_path: str, extract_destination_path: str, index_blob: Any,
-                            build_number: str, private_packs: list, current_commit_hash: str,
-                            index_generation: str, is_private: bool = False):
-    """
-    Upload updated index zip to cloud storage.
-
-    :param index_folder_path: index folder full path.
-    :param extract_destination_path: extract folder full path.
-    :param index_blob: google cloud storage object that represents index.zip blob.
-    :param build_number: circleCI build number, used as an index revision.
-    :param private_packs: List of private packs and their price.
-    :param current_commit_hash: last commit hash of head.
-    :param index_generation: downloaded index generation.
-    :param is_private: Indicates if upload is private.
-    :returns None.
-
-    """
-    with open(os.path.join(index_folder_path, f"{GCPConfig.INDEX_NAME}.json"), "w+") as index_file:
-        index = {
-            'revision': build_number,
-            'modified': datetime.utcnow().strftime(Metadata.DATE_FORMAT),
-            'packs': private_packs,
-            'commit': current_commit_hash
-        }
-        json.dump(index, index_file, indent=4)
-
-    index_zip_name = os.path.basename(index_folder_path)
-    index_zip_path = shutil.make_archive(base_name=index_folder_path, format="zip",
-                                         root_dir=extract_destination_path, base_dir=index_zip_name)
-    try:
-        index_blob.reload()
-        current_index_generation = index_blob.generation
-        index_blob.cache_control = "no-cache,max-age=0"  # disabling caching for index blob
-
-        if is_private or current_index_generation == index_generation:
-            index_blob.upload_from_filename(index_zip_path)
-            logging.success(f"Finished uploading {GCPConfig.INDEX_NAME}.zip to storage.")
-        else:
-            logging.critical(f"Failed in uploading {GCPConfig.INDEX_NAME}, mismatch in index file generation")
-            logging.critical(f"Downloaded index generation: {index_generation}")
-            logging.critical(f"Current index generation: {current_index_generation}")
-            sys.exit(0)
-    except Exception:
-        logging.exception(f"Failed in uploading {GCPConfig.INDEX_NAME}.")
-        sys.exit(1)
-    finally:
-        shutil.rmtree(index_folder_path)
-
-
-def upload_core_packs_config(storage_bucket: Any, build_number: str, index_folder_path: str):
-    """Uploads corepacks.json file configuration to bucket. Corepacks file includes core packs for server installation.
-
-     Args:
-        storage_bucket (google.cloud.storage.bucket.Bucket): gcs bucket where core packs config is uploaded.
-        build_number (str): circleCI build number.
-        index_folder_path (str): The index folder path.
-
-    """
-    core_packs_public_urls = []
-    found_core_packs = set()
-    for pack in os.scandir(index_folder_path):
-        if pack.is_dir() and pack.name in GCPConfig.CORE_PACKS_LIST:
-            pack_metadata_path = os.path.join(index_folder_path, pack.name, Pack.METADATA)
-
-            if not os.path.exists(pack_metadata_path):
-                logging.critical(f"{pack.name} pack {Pack.METADATA} is missing in {GCPConfig.INDEX_NAME}")
-                sys.exit(1)
-
-            with open(pack_metadata_path, 'r') as metadata_file:
-                metadata = json.load(metadata_file)
-
-            pack_current_version = metadata.get('currentVersion', Pack.PACK_INITIAL_VERSION)
-            core_pack_relative_path = os.path.join(GCPConfig.STORAGE_BASE_PATH, pack.name,
-                                                   pack_current_version, f"{pack.name}.zip")
-            core_pack_public_url = os.path.join(GCPConfig.GCS_PUBLIC_URL, storage_bucket.name, core_pack_relative_path)
-
-            if not storage_bucket.blob(core_pack_relative_path).exists():
-                logging.critical(f"{pack.name} pack does not exist under {core_pack_relative_path} path")
-                sys.exit(1)
-
-            core_packs_public_urls.append(core_pack_public_url)
-            found_core_packs.add(pack.name)
-
-    if len(found_core_packs) != len(GCPConfig.CORE_PACKS_LIST):
-        missing_core_packs = set(GCPConfig.CORE_PACKS_LIST) ^ found_core_packs
-        logging.critical(f"Number of defined core packs are: {len(GCPConfig.CORE_PACKS_LIST)}")
-        logging.critical(f"Actual number of found core packs are: {len(found_core_packs)}")
-        logging.critical(f"Missing core packs are: {missing_core_packs}")
-        sys.exit(1)
-
-    # construct core pack data with public gcs urls
-    core_packs_data = {
-        'corePacks': core_packs_public_urls,
-        'buildNumber': build_number
-    }
-    # upload core pack json file to gcs
-    core_packs_config_path = os.path.join(GCPConfig.STORAGE_BASE_PATH, GCPConfig.CORE_PACK_FILE_NAME)
-    blob = storage_bucket.blob(core_packs_config_path)
-    blob.upload_from_string(json.dumps(core_packs_data, indent=4))
-
-    logging.success(f"Finished uploading {GCPConfig.CORE_PACK_FILE_NAME} to storage.")
-
-
-def upload_id_set(storage_bucket: Any, id_set_local_path: str = None):
-    """
-    Uploads the id_set.json artifact to the bucket.
-
-    Args:
-        storage_bucket (google.cloud.storage.bucket.Bucket): gcs bucket where core packs config is uploaded.
-        id_set_local_path: path to the id_set.json file
-    """
-    if not id_set_local_path:
-        logging.info("Skipping upload of id set to gcs.")
-        return
-
-    id_set_gcs_path = os.path.join(GCPConfig.STORAGE_CONTENT_PATH, 'id_set.json')
-    blob = storage_bucket.blob(id_set_gcs_path)
-
-    with open(id_set_local_path, mode='r') as f:
-        blob.upload_from_file(f)
-    logging.success("Finished uploading id_set.json to storage.")
 
 
 def get_private_packs(private_index_path: str, pack_names: set, is_private_build: bool,
@@ -499,171 +166,6 @@ def update_index_with_priced_packs(private_storage_bucket: Any, extract_destinat
         return private_packs, private_index_path, private_index_blob
 
 
-def _build_summary_table(packs_input_list: list, include_pack_status: bool = False) -> Any:
-    """Build summary table from pack list
-
-    Args:
-        packs_input_list (list): list of Packs
-
-    Returns:
-        PrettyTable: table with upload result of packs.
-
-    """
-    table_fields = ["Index", "Pack ID", "Pack Display Name", "Latest Version", "Status",
-                    "Pack Bucket URL"] if include_pack_status \
-        else ["Index", "Pack ID", "Pack Display Name", "Latest Version", "Pack Bucket URL"]
-    table = prettytable.PrettyTable()
-    table.field_names = table_fields
-
-    for index, pack in enumerate(packs_input_list, start=1):
-        pack_status_message = PackStatus[pack.status].value
-        row = [index, pack.name, pack.display_name, pack.latest_version, pack_status_message,
-               pack.bucket_url] if include_pack_status \
-            else [index, pack.name, pack.display_name, pack.latest_version, pack.bucket_url]
-        table.add_row(row)
-
-    return table
-
-
-def build_summary_table_md(packs_input_list: list, include_pack_status: bool = False) -> str:
-    """Build markdown summary table from pack list
-
-    Args:
-        packs_input_list (list): list of Packs
-        include_pack_status (bool): whether pack includes status
-
-    Returns:
-        Markdown table: table with upload result of packs.
-
-    """
-    table_fields = ["Index", "Pack ID", "Pack Display Name", "Latest Version", "Status",
-                    "Pack Bucket URL"] if include_pack_status \
-        else ["Index", "Pack ID", "Pack Display Name", "Latest Version", "Pack Bucket URL"]
-
-    table = ['|', '|']
-
-    for key in table_fields:
-        table[0] = f'{table[0]} {key} |'
-        table[1] = f'{table[1]} :- |'
-
-    for index, pack in enumerate(packs_input_list):
-        pack_status_message = PackStatus[pack.status].value if include_pack_status else ''
-
-        row = [index, pack.name, pack.display_name, pack.latest_version, pack_status_message,
-               pack.bucket_url] if include_pack_status \
-            else [index, pack.name, pack.display_name, pack.latest_version, pack.bucket_url]
-
-        row_hr = '|'
-        for _value in row:
-            row_hr = f'{row_hr} {_value}|'
-        table.append(row_hr)
-
-    return '\n'.join(table)
-
-
-def load_json(file_path: str) -> dict:
-    """ Reads and loads json file.
-
-    Args:
-        file_path (str): full path to json file.
-
-    Returns:
-        dict: loaded json file.
-
-    """
-    if file_path:
-        with open(file_path, 'r') as json_file:
-            result = json.load(json_file)
-    else:
-        result = {}
-    return result
-
-
-def get_content_git_client(content_repo_path: str):
-    """ Initializes content repo client.
-
-    Args:
-        content_repo_path (str): content repo full path
-
-    Returns:
-        git.repo.base.Repo: content repo object.
-
-    """
-    return git.Repo(content_repo_path)
-
-
-def get_recent_commits_data(content_repo: Any):
-    """ Returns recent commits hashes (of head and remote master)
-
-    Args:
-        content_repo (git.repo.base.Repo): content repo object.
-
-    Returns:
-        str: last commit hash of head.
-        str: previous commit of origin/master (origin/master~1)
-    """
-    return content_repo.head.commit.hexsha, content_repo.commit('origin/master~1').hexsha
-
-
-def check_if_index_is_updated(index_folder_path: str, content_repo: Any, current_commit_hash: str,
-                              remote_previous_commit_hash: str, storage_bucket: Any):
-    """ Checks stored at index.json commit hash and compares it to current commit hash. In case no
-    packs folders were added/modified/deleted, all other steps are not performed.
-
-    Args:
-        index_folder_path (str): index folder full path.
-        content_repo (git.repo.base.Repo): content repo object.
-        current_commit_hash (str): last commit hash of head.
-        remote_previous_commit_hash (str): previous commit of origin/master (origin/master~1)
-        storage_bucket: public storage bucket.
-
-    """
-    skipping_build_task_message = "Skipping Upload Packs To Marketplace Storage Step."
-
-    try:
-        if storage_bucket.name != GCPConfig.PRODUCTION_BUCKET:
-            logging.info("Skipping index update check in non production bucket")
-            return
-
-        if not os.path.exists(os.path.join(index_folder_path, f"{GCPConfig.INDEX_NAME}.json")):
-            # will happen only in init bucket run
-            logging.warning(f"{GCPConfig.INDEX_NAME}.json not found in {GCPConfig.INDEX_NAME} folder")
-            return
-
-        with open(os.path.join(index_folder_path, f"{GCPConfig.INDEX_NAME}.json")) as index_file:
-            index_json = json.load(index_file)
-
-        index_commit_hash = index_json.get('commit', remote_previous_commit_hash)
-
-        try:
-            index_commit = content_repo.commit(index_commit_hash)
-        except Exception:
-            # not updated build will receive this exception because it is missing more updated commit
-            logging.exception(f"Index is already updated. {skipping_build_task_message}")
-            sys.exit()
-
-        current_commit = content_repo.commit(current_commit_hash)
-
-        if current_commit.committed_datetime <= index_commit.committed_datetime:
-            logging.warning(f"Current commit {current_commit.hexsha} committed time: {current_commit.committed_datetime}")
-            logging.warning(f"Index commit {index_commit.hexsha} committed time: {index_commit.committed_datetime}")
-            logging.warning("Index is already updated.")
-            logging.warning(skipping_build_task_message)
-            sys.exit()
-
-        for changed_file in current_commit.diff(index_commit):
-            if changed_file.a_path.startswith(PACKS_FOLDER):
-                logging.info(f"Found changed packs between index commit {index_commit.hexsha} and {current_commit.hexsha}")
-                break
-        else:
-            logging.warning(f"No changes found between index commit {index_commit.hexsha} and {current_commit.hexsha}")
-            logging.warning(skipping_build_task_message)
-            sys.exit()
-    except Exception:
-        logging.exception("Failed in checking status of index")
-        sys.exit(1)
-
-
 def should_upload_core_packs(storage_bucket_name: str) -> bool:
     """
     Indicates if the core packs should be updated.
@@ -674,96 +176,6 @@ def should_upload_core_packs(storage_bucket_name: str) -> bool:
     is_private_storage_bucket = (storage_bucket_name != GCPConfig.PRODUCTION_PRIVATE_BUCKET)
     is_private_ci_bucket = (storage_bucket_name != GCPConfig.CI_PRIVATE_BUCKET)
     return not (is_private_storage_bucket or is_private_ci_bucket)
-
-
-def print_packs_summary(packs_list: list):
-    """Prints summary of packs uploaded to gcs.
-
-    Args:
-        packs_list (list): list of initialized packs.
-
-    """
-    successful_packs = [pack for pack in packs_list if pack.status == PackStatus.SUCCESS.name]
-    skipped_packs = [pack for pack in packs_list if
-                     pack.status == PackStatus.PACK_ALREADY_EXISTS.name
-                     or pack.status == PackStatus.PACK_IS_NOT_UPDATED_IN_RUNNING_BUILD.name
-                     or pack.status == PackStatus.FAILED_DETECTING_MODIFIED_FILES.name]
-    failed_packs = [pack for pack in packs_list if pack not in successful_packs and pack not in skipped_packs]
-
-    logging.info(
-        f"""------------------------------------------ Packs Upload Summary ------------------------------------------
-        Total number of packs: {len(packs_list)}
-        ----------------------------------------------------------------------------------------------------------""")
-    if successful_packs:
-        successful_packs_table = _build_summary_table(successful_packs)
-        logging.success(f"Number of successful uploaded packs: {len(successful_packs)}")
-        logging.success(f"Uploaded packs:\n{successful_packs_table}")
-        with open('pack_list.txt', 'w') as f:
-            f.write(successful_packs_table.get_string())
-    if skipped_packs:
-        skipped_packs_table = _build_summary_table(skipped_packs)
-        logging.warning(f"Number of skipped packs: {len(skipped_packs)}\nSkipped packs:\n{skipped_packs_table}")
-    if failed_packs:
-        failed_packs_table = _build_summary_table(failed_packs, include_pack_status=True)
-        logging.critical(f"Number of failed packs: {len(failed_packs)}\nFailed packs:\n{failed_packs_table}")
-        sys.exit(1)
-
-    # for external pull requests -  when there is no failed packs, add the build summary to the pull request
-    branch_name = os.environ.get('CIRCLE_BRANCH')
-    if branch_name and branch_name.startswith('pull/'):
-        successful_packs_table = build_summary_table_md(successful_packs)
-
-        build_num = os.environ['CIRCLE_BUILD_NUM']
-
-        bucket_path = f'https://console.cloud.google.com/storage/browser/' \
-            f'marketplace-ci-build/content/builds/{branch_name}/{build_num}'
-
-        pr_comment = f'Number of successful uploaded packs: {len(successful_packs)}\n' \
-            f'Uploaded packs:\n{successful_packs_table}\n\n' \
-            f'Browse to the build bucket with this address:\n{bucket_path}'
-
-        add_pr_comment(pr_comment)
-
-
-def add_pr_comment(comment: str):
-    """Add comment to the pull request.
-
-    Args:
-        comment (string): The comment text.
-
-    """
-    token = os.environ['CONTENT_GITHUB_TOKEN']
-    branch_name = os.environ['CIRCLE_BRANCH']
-    sha1 = os.environ['CIRCLE_SHA1']
-
-    query = f'?q={sha1}+repo:demisto/content+is:pr+is:open+head:{branch_name}+is:open'
-    url = 'https://api.github.com/search/issues'
-    headers = {'Authorization': 'Bearer ' + token}
-    try:
-        res = requests.get(url + query, headers=headers, verify=False)
-        res = handle_github_response(res)
-        if res and res.get('total_count', 0) == 1:
-            issue_url = res['items'][0].get('comments_url') if res.get('items', []) else None
-            if issue_url:
-                res = requests.post(issue_url, json={'body': comment}, headers=headers, verify=False)
-                handle_github_response(res)
-        else:
-            logging.warning(
-                f'Add pull request comment failed: There is more then one open pull request for branch {branch_name}.')
-    except Exception:
-        logging.exception('Add pull request comment failed.')
-
-
-def handle_github_response(response: json) -> dict:
-    """
-    Handles the response from the GitHub server after making a request.
-    :param response: Response from the server.
-    :return: The returned response.
-    """
-    res_dict = response.json()
-    if not res_dict.get('ok'):
-        logging.warning(f'Add pull request comment failed: {res_dict.get("message")}')
-    return res_dict
 
 
 def create_and_upload_marketplace_pack(upload_config: Any, pack: Any, storage_bucket: Any, index_folder_path: str,
